requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
faiss-cpu
PyPDF2
langchain_google_genai
langchain-community
pandas
pdfplumber
plotly
seaborn
matplotlib
sentence-transformers
fpdf
transformers
kaleido